more reattach tests
