+++ internal/e2e-js/playwright.config.ts
@@ -37,7 +37,10 @@ const reattachTests = [
 ]
 const callfabricTests = [
   'address.spec.ts',
+  'buildVideoElement.spec.ts',
   'conversation.spec.ts',
+  'reattach.spec.ts',
+  'reattachWithDial.spec.ts',
   'relayApp.spec.ts',
   'swml.spec.ts',
   'videoRoom.spec.ts',
